refactor: 멤버 API URL 변경 #33
- profiles -> members
