+++ src/domain/profiles/services/memberService.ts
@@ -2,10 +2,12 @@ import apiService from "../../../common/apis/apiService";
 import type { MessageResponse } from "../../../common/types/messageResponse";
 import { NotifyPrayerRequest } from "../../notifications/types/request/notifyPrayerRequest";
 
-export const profileService = {
+export const memberService = {
   // 사용자 프로필 단일 정보 조회
-  fetch: async () => {
-    const response = await apiService.get<FetchProfileResponse>(`/profiles/me`);
+  fetchProfile: async () => {
+    const response = await apiService.get<FetchProfileResponse>(
+      `/members/profiles`
+    );
     return response.data;
   },
 };
